# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JARVIS",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/workflows": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Create Workflow",
        "operationId": "create_workflow_v1_workflows_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "List Workflows",
        "operationId": "list_workflows_v1_workflows_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Get Workflow",
        "operationId": "get_workflow_v1_workflows__workflow_id__get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "workflows"
        ],
        "summary": "Update Workflow",
        "operationId": "update_workflow_v1_workflows__workflow_id__patch",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workflows"
        ],
        "summary": "Delete Workflow",
        "operationId": "delete_workflow_v1_workflows__workflow_id__delete",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}/run": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Run Workflow",
        "description": "Blocking run \u2014 enqueue with kind=\"workflow\" (D41), then wait for the\nsegment to end. A pause INSIDE an agent node (S10 composition, ADR 0015\n\u00a77) ends the segment like an agent pause: the route returns the\nawaiting_input row and the client resumes with\nPOST /executions/{id}/resume.",
        "operationId": "run_workflow_v1_workflows__workflow_id__run_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}/stream": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Stream Workflow",
        "description": "SSE run: enqueue with kind=\"workflow\", then frame every event the\nworker writes \u2014 including `node.*` events stamped with the executing\nnode's node_id (D43). Resume with `Last-Event-ID` plus `run_id`, exactly\nlike the agents stream.",
        "operationId": "stream_workflow_v1_workflows__workflow_id__stream_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}/versions/{version}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Get Workflow Version",
        "operationId": "get_workflow_version_v1_workflows__workflow_id__versions__version__get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowVersion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunRequest": {
        "properties": {
          "input": {
            "type": "string",
            "minLength": 1,
            "title": "Input"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "variables": {
            "additionalProperties": true,
            "type": "object",
            "title": "Variables"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "input"
        ],
        "title": "RunRequest",
        "description": "Body for /run and /stream. `run_id` is stream-resume only: attach to\nan existing run's stream instead of starting a new one."
      },
      "RunResult": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "awaiting_input",
              "succeeded",
              "failed",
              "cancelled",
              "timed_out"
            ],
            "title": "Status"
          },
          "input": {
            "type": "string",
            "title": "Input",
            "default": ""
          },
          "agent_version_id": {
            "type": "string",
            "title": "Agent Version Id",
            "default": ""
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "default": ""
          },
          "final_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Message"
          },
          "total_usage": {
            "$ref": "#/components/schemas/Usage"
          },
          "iterations": {
            "type": "integer",
            "title": "Iterations",
            "default": 0
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Kind"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "event_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Cursor"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "agent_id",
          "status"
        ],
        "title": "RunResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionSummary": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "id",
          "label",
          "created_at"
        ],
        "title": "VersionSummary"
      },
...
      "WorkflowDefinition": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/WorkflowNode"
            },
            "type": "array",
            "minItems": 1,
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/WorkflowEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "start_node_id": {
            "type": "string",
            "minLength": 1,
            "title": "Start Node Id"
          },
          "max_node_executions": {
            "type": "integer",
            "maximum": 128.0,
            "minimum": 1.0,
            "title": "Max Node Executions",
            "default": 24
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "nodes",
          "start_node_id"
        ],
        "title": "WorkflowDefinition"
      },
      "WorkflowDetail": {
        "properties": {
          "definition": {
            "$ref": "#/components/schemas/WorkflowDefinition"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionSummary"
            },
            "type": "array",
            "title": "Versions"
          },
          "lints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Lints"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "definition",
          "versions"
        ],
        "title": "WorkflowDetail"
      },
      "WorkflowEdge": {
        "properties": {
          "from_node": {
            "type": "string",
            "minLength": 1,
            "title": "From Node"
          },
          "to_node": {
            "type": "string",
            "minLength": 1,
            "title": "To Node"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "from_node",
          "to_node"
        ],
        "title": "WorkflowEdge"
      },
      "WorkflowList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkflowDefinition"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "WorkflowList"
      },
      "WorkflowNode": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "agent",
              "tool",
              "condition"
            ],
            "title": "Type"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentNodeConfig"
              },
              {
                "$ref": "#/components/schemas/ToolNodeConfig"
              },
              {
                "$ref": "#/components/schemas/ConditionNodeConfig"
              }
            ],
            "title": "Config"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "config"
        ],
        "title": "WorkflowNode"
      },
      "WorkflowUpsertRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/WorkflowNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/WorkflowEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "start_node_id": {
            "type": "string",
            "minLength": 1,
            "title": "Start Node Id"
          },
          "max_node_executions": {
            "type": "integer",
            "maximum": 128.0,
            "minimum": 1.0,
            "title": "Max Node Executions",
            "default": 24
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "nodes",
          "start_node_id"
        ],
        "title": "WorkflowUpsertRequest",
        "description": "Create/patch payload for a workflow (S6, ADR 0015 \u00a76) \u2014 the graph the\nclient authored. Node configs validate against the typed per-type\nconfigs; `agent_version_id` stays None on drafts (publish pins, D42)."
      },
      "WorkflowVersion": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "snapshot": {
            "$ref": "#/components/schemas/WorkflowDefinition"
          },
          "label": {
            "type": "string",
            "title": "Label",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "workflow_id",
          "version",
          "snapshot"
        ],
        "title": "WorkflowVersion",
        "description": "Immutable, append-only snapshot of a WorkflowDefinition \u2014 the same\nD1 replay argument as AgentVersion, one level up."
      }
    }
  }
}